requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.13
sqlalchemy==2.0.23
pymongo==4.6.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
PyJWT==2.8.0
httpx==0.25.2
sqlmodel==0.0.14
pymysql==1.1.0
google-api-python-client==2.108.0 